requirement.txt
Flask
psycopg2-binary
SQLAlchemy
flask-restful
flask_sqlalchemy
python-dotenv
jsonschema
redis
WSGIserver